pyproject.toml
[project]
name = "building-energy-3d"
version = "0.1.0"
description = "3D Building Energy Simulation Map Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "psycopg2-binary>=2.9",
    "geoalchemy2>=0.15",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "httpx>=0.27",
    "pandas>=2.1",
    "geopandas>=0.14",
    "shapely>=2.0",
    "trimesh>=4.0",
    "numpy>=1.26",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "PublicDataReader>=1.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "black>=24.0",
]
sim = [
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "openstudio>=3.7",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
